move package for consensus entity
