+++ confignode/src/main/java/org/apache/iotdb/confignode/physical/sys/RegisterDataNodePlan.java
@@ -19,7 +19,7 @@
 package org.apache.iotdb.confignode.physical.sys;
 
 import org.apache.iotdb.commons.partition.DataNodeLocation;
-import org.apache.iotdb.commons.partition.Endpoint;
+import org.apache.iotdb.commons.cluster.Endpoint;
 import org.apache.iotdb.confignode.physical.PhysicalPlan;
 import org.apache.iotdb.confignode.physical.PhysicalPlanType;
