+++ consensus/src/main/java/org/apache/iotdb/consensus/ratis/RatisConsensus.java
@@ -19,8 +19,8 @@
 
 package org.apache.iotdb.consensus.ratis;
 
-import org.apache.iotdb.commons.partition.ConsensusGroupId;
-import org.apache.iotdb.commons.partition.Endpoint;
+import org.apache.iotdb.commons.consensus.ConsensusGroupId;
+import org.apache.iotdb.commons.cluster.Endpoint;
 import org.apache.iotdb.consensus.IConsensus;
 import org.apache.iotdb.consensus.common.DataSet;
 import org.apache.iotdb.consensus.common.Peer;
