+++ node-commons/src/main/java/org/apache/iotdb/commons/partition/RegionReplicaSet.java
@@ -18,6 +18,8 @@
  */
 package org.apache.iotdb.commons.partition;
 
+import org.apache.iotdb.commons.consensus.ConsensusGroupId;
+
 import java.util.List;
 
 public class RegionReplicaSet {
